+++ example/lib/src/biz/biz1/biz1.route.dart
@@ -6,6 +6,7 @@
 import 'package:flutter_thrio/flutter_thrio.dart';
 
 import 'flutter1/flutter1.route.dart';
+import 'flutter11/flutter11.route.dart';
 import 'flutter3/flutter3.route.dart';
 import 'flutter5/flutter5.route.dart';
 import 'flutter7/flutter7.route.dart';
@@ -28,6 +29,8 @@ class Biz1Route extends NavigatorRouteNode {
   late final flutter7 = Flutter7Route(this);
 
   late final flutter9 = Flutter9Route(this);
+
+  late final flutter11 = Flutter11Route(this);
 
   @override
   String get name => 'biz1';