+++ example/lib/src/biz/biz1/flutter11/flutter11.state.dart
@@ -1,0 +1,7 @@
+// Copyright (c) 2023 foxsofter.
+//
+// ignore_for_file: avoid_as
+
+part of 'flutter11.page.dart';
+
+extension Flutter11 on _Flutter11PageState {}