+++ example/lib/src/biz/biz1/module.dart
@@ -6,6 +6,7 @@
 import 'package:flutter_thrio/flutter_thrio.dart';
 
 import 'flutter1/module.dart' as flutter1;
+import 'flutter11/module.dart' as flutter11;
 import 'flutter3/module.dart' as flutter3;
 import 'flutter5/module.dart' as flutter5;
 import 'flutter7/module.dart' as flutter7;
@@ -22,6 +23,7 @@ class Module with ThrioModule, ModuleParamScheme {
     registerModule(flutter5.Module(), moduleContext);
     registerModule(flutter7.Module(), moduleContext);
     registerModule(flutter9.Module(), moduleContext);
+    registerModule(flutter11.Module(), moduleContext);
   }
 
   @override